modal style

+++ pages/api/products/[id].ts
@@ -13,31 +13,32 @@ export default async function handler(
 
   await dbConnect();
 
-  switch (method) {
-    case 'PUT':
-      try {
-        const product = await Product.findByIdAndUpdate(id, req.body, {
-          new: true,
-        });
-        res.status(200).json(product);
-      } catch (err) {
-        res.status(500).json(err);
-      }
-      break;
-    case 'DELETE':
-      try {
-        await Product.findByIdAndDelete(id);
-        res.status(200).json('The product has been deleted!');
-      } catch (err) {
-        res.status(500).json(err);
-      }
-      break;
-    default:
-      try {
-        const product = await Product.findById(id);
-        res.status(200).json(product);
-      } catch (err) {
-        res.status(500).json(err);
-      }
+  if (method === 'PUT') {
+    try {
+      const product = await Product.findByIdAndUpdate(id, req.body, {
+        new: true,
+      });
+      res.status(200).json(product);
+    } catch (err) {
+      res.status(500).json(err);
+    }
+  }
+
+  if (method === 'DELETE') {
+    try {
+      await Product.findByIdAndDelete(id);
+      res.status(200).json('The product has been deleted!');
+    } catch (err) {
+      res.status(500).json(err);
+    }
+  }
+
+  if (method === 'GET') {
+    try {
+      const product = await Product.findById(id);
+      res.status(200).json(product);
+    } catch (err) {
+      res.status(500).json(err);
+    }
   }
 }
